// extensions/ns-nl-search/src/api/types.ts
export interface Station {
  /**
   *
   * @type {string}
   * @memberof Station
   */
  UICCode: string;
  /**
   *
   * @type {string}
   * @memberof Station
   */
  stationType: string;
  /**
   *
   * @type {string}
   * @memberof Station
   */
  EVACode?: string;
  /**
   *
   * @type {number}
   * @memberof Station
   */
  cdCode?: number;
  /**
   *
   * @type {string}
   * @memberof Station
   */
  code?: string;
  /**
   *
   * @type {Array<Track>}
   * @memberof Station
   */
  sporen: Array<Track>;
  /**
   *
   * @type {Array<string>}
   * @memberof Station
   */
  synoniemen: Array<string>;
  /**
   *
   * @type {boolean}
   * @memberof Station
   */
  heeftFaciliteiten: boolean;
  /**
   *
   * @type {boolean}
   * @memberof Station
   */
  heeftVertrektijden: boolean;
  /**
   *
   * @type {boolean}
   * @memberof Station
   */
  heeftReisassistentie: boolean;
  /**
   *
   * @type {StationsNamen}
   * @memberof Station
   */
  namen?: StationsNamen;
  /**
   *
   * @type {string}
   * @memberof Station
   */
  land?: string;
  /**
   *
   * @type {number}
   * @memberof Station
   */
  lat?: number;
  /**
   *
   * @type {number}
   * @memberof Station
   */
  lng?: number;
  /**
   *
   * @type {number}
   * @memberof Station
   */
  radius?: number;
  /**
   *
   * @type {number}
   * @memberof Station
   */
  naderenRadius?: number;
  /**
   *
   * @type {number}
   * @memberof Station
   */
  distance?: number;
  /**
   *
   * @type {string}
   * @memberof Station
   */
  ingangsDatum?: string;
  /**
   *
   * @type {string}
   * @memberof Station
   */
  eindDatum?: string;
  /**
   *
   * @type {NearbyMeLocationId}
   * @memberof Station
   */
  nearbyMeLocationId?: NearbyMeLocationId;
}
// ...
export interface ArrivalOrDeparture {
  /**
   *
   * @type {Product}
   * @memberof ArrivalOrDeparture
   */
  product: Product;
  /**
   *
   * @type {Station}
   * @memberof ArrivalOrDeparture
   */
  origin?: Station;
  /**
   *
   * @type {Station}
   * @memberof ArrivalOrDeparture
   */
  destination?: Station;
  /**
   *
   * @type {Date}
   * @memberof ArrivalOrDeparture
   */
  plannedTime?: Date;
  /**
   *
   * @type {Date}
   * @memberof ArrivalOrDeparture
   */
  actualTime?: Date;
  /**
   *
   * @type {number}
   * @memberof ArrivalOrDeparture
   */
  delayInSeconds?: number;
  /**
   *
   * @type {string}
   * @memberof ArrivalOrDeparture
   */
  plannedTrack?: string;
  /**
   *
   * @type {string}
   * @memberof ArrivalOrDeparture
   */
  actualTrack?: string;
  /**
   *
   * @type {boolean}
   * @memberof ArrivalOrDeparture
   */
  cancelled: boolean;
  /**
   *
   * @type {number}
   * @memberof ArrivalOrDeparture
   */
  punctuality?: number;
  /**
   *
   * @type {string}
   * @memberof ArrivalOrDeparture
   */
  crowdForecast: ArrivalOrDeparture.CrowdForecastEnum;
  /**
   *
   * @type {string}
   * @memberof ArrivalOrDeparture
   */
  shorterStockClassification?: ArrivalOrDeparture.ShorterStockClassificationEnum;
  /**
   *
   * @type {Array<string>}
   * @memberof ArrivalOrDeparture
   */
  stockIdentifiers?: Array<string>;
}

/**
 * @export
 * @namespace ArrivalOrDeparture
 */
export namespace ArrivalOrDeparture {
  /**
   * @export
   * @enum {string}
   */
  export enum CrowdForecastEnum {
    UNKNOWN = <any>"UNKNOWN",
    LOW = <any>"LOW",
    MEDIUM = <any>"MEDIUM",
    HIGH = <any>"HIGH",
  }
  /**
   * @export
   * @enum {string}
   */
  export enum ShorterStockClassificationEnum {
    BUSY = <any>"BUSY",
    EXTRABUSY = <any>"EXTRA_BUSY",
  }
}
// ...
  export enum StatusEnum {
    REGISTRATIONPOSSIBLE = <any>"REGISTRATION_POSSIBLE",
    NOTAVAILABLE = <any>"NOT_AVAILABLE",
    DATEINPAST = <any>"DATE_IN_PAST",
    DATETOOFARFUTURE = <any>"DATE_TOO_FAR_FUTURE",
    NOTNECESSARYOTHEROPERATOR = <any>"NOT_NECESSARY_OTHER_OPERATOR",
    UNKNOWN = <any>"UNKNOWN",
  }
// ...
export namespace Leg {
  /**
   * @export
   * @enum {string}
   */
  export enum TravelTypeEnum {
    PUBLICTRANSIT = <any>"PUBLIC_TRANSIT",
    WALK = <any>"WALK",
    TRANSFER = <any>"TRANSFER",
    BIKE = <any>"BIKE",
    CAR = <any>"CAR",
    KISS = <any>"KISS",
    TAXI = <any>"TAXI",
    UNKNOWN = <any>"UNKNOWN",
  }
  /**
   * @export
   * @enum {string}
   */
  export enum CrowdForecastEnum {
    UNKNOWN = <any>"UNKNOWN",
    LOW = <any>"LOW",
    MEDIUM = <any>"MEDIUM",
    HIGH = <any>"HIGH",
  }
  /**
   * @export
   * @enum {string}
   */
  export enum ShorterStockClassificationEnum {
    BUSY = <any>"BUSY",
    EXTRABUSY = <any>"EXTRA_BUSY",
  }
}
// ...
export namespace Message {
  /**
   * @export
   * @enum {string}
   */
  export enum TypeEnum {
    MAINTENANCE = <any>"MAINTENANCE",
    DISRUPTION = <any>"DISRUPTION",
    SHORTENED = <any>"SHORTENED",
  }
  /**
   * @export
   * @enum {string}
   */
  export enum PhaseEnum {
    _1A = <any>"PHASE_1A",
    _1B = <any>"PHASE_1B",
    _2 = <any>"PHASE_2",
    _3 = <any>"PHASE_3",
    _4 = <any>"PHASE_4",
    _5 = <any>"PHASE_5",
  }
}
// ...
export interface PrimaryMessage {
  /**
   *
   * @type {string}
   * @memberof PrimaryMessage
   */
  title: string;
  /**
   *
   * @type {NesProperties}
   * @memberof PrimaryMessage
   */
  nesProperties: NesProperties;
  /**
   *
   * @type {Message}
   * @memberof PrimaryMessage
   */
  message?: Message;
  /**
   *
   * @type {string}
   * @memberof PrimaryMessage
   */
  type: PrimaryMessage.TypeEnum;
}

/**
 * @export
 * @namespace PrimaryMessage
 */
export namespace PrimaryMessage {
  /**
   * @export
   * @enum {string}
   */
  export enum TypeEnum {
    TRIPCANCELLED = <any>"TRIP_CANCELLED",
    LEGCANCELLED = <any>"LEG_CANCELLED",
    LEGTRANSFERIMPOSSIBLE = <any>"LEG_TRANSFER_IMPOSSIBLE",
    ALTERNATIVETRANSPORT = <any>"ALTERNATIVE_TRANSPORT",
    DISRUPTION = <any>"DISRUPTION",
    MAINTENANCE = <any>"MAINTENANCE",
    REPLACEMENT = <any>"REPLACEMENT",
    ADDITIONAL = <any>"ADDITIONAL",
    SHORTENEDTRAIN = <any>"SHORTENED_TRAIN",
  }
}
/**
 *
 * @export
 * @interface Product
 */
export interface Product {
  /**
   *
   * @type {string}
   * @memberof Product
   */
  number?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  categoryCode?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  shortCategoryName?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  longCategoryName?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  operatorCode?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  operatorName?: string;
  /**
   *
   * @type {number}
   * @memberof Product
   */
  operatorAdministrativeCode?: number;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  type: Product.TypeEnum;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  displayName?: string;
  /**
   *
   * @type {NesProperties}
   * @memberof Product
   */
  nameNesProperties: NesProperties;
  /**
   *
   * @type {NesProperties}
   * @memberof Product
   */
  iconNesProperties?: NesProperties;
  /**
   * notes is a List of Lists because this allows frontend clients to render the elements of every nested List on the same row
   * @type {Array<Array<Note>>}
   * @memberof Product
   */
  notes: Array<Array<Note>>;
  /**
   *
   * @type {NesProperties}
   * @memberof Product
   */
  nesProperties?: NesProperties;
}

/**
 * @export
 * @namespace Product
 */
export namespace Product {
  /**
   * @export
   * @enum {string}
   */
  export enum TypeEnum {
    TRAIN = <any>"TRAIN",
    BUS = <any>"BUS",
    TRAM = <any>"TRAM",
    METRO = <any>"METRO",
    FERRY = <any>"FERRY",
    WALK = <any>"WALK",
    BIKE = <any>"BIKE",
    CAR = <any>"CAR",
    TAXI = <any>"TAXI",
    SHAREDMODALITY = <any>"SHARED_MODALITY",
    UNKNOWN = <any>"UNKNOWN",
  }
}
// ...
export interface Trip {
  /**
   * Unique identifier for this trip
   * @type {string}
   * @memberof Trip
   */
  uid: string;
  /**
   * Reconstruction context for this trip. Can be used to reconstruct this exact trip with the v3/trips/trip endpoint
   * @type {string}
   * @memberof Trip
   */
  ctxRecon: string;
  /**
   * Reconstruction context for this trip set by the travel advice source system. Can be used to interact follow-up requests with the source system, like booking a trip with travel assistance (NSRA). Can be null if the source doesn't provide a ctxRecon
   * @type {string}
   * @memberof Trip
   */
  sourceCtxRecon?: string;
  /**
   * Planned duration of this trip in minutes
   * @type {number}
   * @memberof Trip
   */
  plannedDurationInMinutes?: number;
  /**
   * Actual duration of this trip in minutes, or the planned duration if no realtime information about this trip is available.
   * @type {number}
   * @memberof Trip
   */
  actualDurationInMinutes?: number;
  /**
   * Number of public transit transfers
   * @type {number}
   * @memberof Trip
   */
  transfers: number;
  /**
   * Status of this trip
   * @type {string}
   * @memberof Trip
   */
  status: Trip.StatusEnum;
  /**
   *
   * @type {PrimaryMessage}
   * @memberof Trip
   */
  primaryMessage?: PrimaryMessage;
  /**
   * List of messages regarding maintenance or disruption that influences this trip.
   * @type {Array<Message>}
   * @memberof Trip
   */
  messages?: Array<Message>;
  /**
   *
   * @type {Array<Leg>}
   * @memberof Trip
   */
  legs: Array<Leg>;
  /**
   *
   * @type {Array<Coordinate>}
   * @memberof Trip
   */
  overviewPolyLine?: Array<Coordinate>;
  /**
   *
   * @type {string}
   * @memberof Trip
   */
  crowdForecast?: Trip.CrowdForecastEnum;
  /**
   *
   * @type {number}
   * @memberof Trip
   */
  punctuality?: number;
  /**
   * Whether or not this trip is regarded the best possible option of all returned trips
   * @type {boolean}
   * @memberof Trip
   */
  optimal: boolean;
  /**
   *
   * @type {FareRoute}
   * @memberof Trip
   */
  fareRoute?: FareRoute;
  /**
   *
   * @type {Array<TripSalesFare>}
   * @memberof Trip
   */
  fares?: Array<TripSalesFare>;
  /**
   *
   * @type {Array<FareLeg>}
   * @memberof Trip
   */
  fareLegs?: Array<FareLeg>;
  /**
   *
   * @type {TripTravelFare}
   * @memberof Trip
   */
  productFare?: TripTravelFare;
  /**
   *
   * @type {TripFareOptions}
   * @memberof Trip
   */
  fareOptions?: TripFareOptions;
  /**
   *
   * @type {Link}
   * @memberof Trip
   */
  bookingUrl?: Link;
  /**
   *
   * @type {NsiLink}
   * @memberof Trip
   */
  nsiLink?: NsiLink;
  /**
   *
   * @type {string}
   * @memberof Trip
   */
  type: Trip.TypeEnum;
  /**
   *
   * @type {Link}
   * @memberof Trip
   */
  shareUrl?: Link;
  /**
   *
   * @type {boolean}
   * @memberof Trip
   */
  realtime: boolean;
  /**
   *
   * @type {TravelAssistanceInfo}
   * @memberof Trip
   */
  travelAssistanceInfo?: TravelAssistanceInfo;
  /**
   *
   * @type {string}
   * @memberof Trip
   */
  routeId?: string;
  /**
   *
   * @type {JourneyRegistrationParameters}
   * @memberof Trip
   */
  registerJourney?: JourneyRegistrationParameters;
  /**
   *
   * @type {Eco}
   * @memberof Trip
   */
  eco?: Eco;
  /**
   *
   * @type {Array<ModalityListItem>}
   * @memberof Trip
   */
  modalityListItems?: Array<ModalityListItem>;
  /**
   *
   * @type {Array<LabelListItem>}
   * @memberof Trip
   */
  labelListItems?: Array<LabelListItem>;
}

/**
 * @export
 * @namespace Trip
 */
export namespace Trip {
  /**
   * @export
   * @enum {string}
   */
  export enum StatusEnum {
    CANCELLED = <any>"CANCELLED",
    CHANGENOTPOSSIBLE = <any>"CHANGE_NOT_POSSIBLE",
    ALTERNATIVETRANSPORT = <any>"ALTERNATIVE_TRANSPORT",
    DISRUPTION = <any>"DISRUPTION",
    MAINTENANCE = <any>"MAINTENANCE",
    UNCERTAIN = <any>"UNCERTAIN",
    REPLACEMENT = <any>"REPLACEMENT",
    ADDITIONAL = <any>"ADDITIONAL",
    SPECIAL = <any>"SPECIAL",
    NORMAL = <any>"NORMAL",
  }
  /**
   * @export
   * @enum {string}
   */
  export enum CrowdForecastEnum {
    UNKNOWN = <any>"UNKNOWN",
    LOW = <any>"LOW",
    MEDIUM = <any>"MEDIUM",
    HIGH = <any>"HIGH",
  }
  /**
   * @export
   * @enum {string}
   */
  export enum TypeEnum {
    NS = <any>"NS",
    NSACCESSIBLE = <any>"NS_ACCESSIBLE",
    NEGENTWEE = <any>"NEGENTWEE",
    GOOGLE = <any>"GOOGLE",
    PAS = <any>"PAS",
  }
}